web/requirements.txt
django-grappelli
